sales_bot/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiohttp==3.9.3
python-jose==3.3.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.9 